+++ packages/dialtone-css/postcss/constants.cjs
@@ -249,6 +249,23 @@ module.exports = {
       'ExtraExtraLarge',
 
     ].join('|'),
+    TEXT_DECORATION: [
+      'dotted',
+      'line-through',
+      'none',
+      'underline',
+      'unset',
+    ].join('|'),
+    OPACITY_VARIATIONS: [
+      '0',
+      '10',
+      '25',
+      '50',
+      '75',
+      '90',
+      '95',
+      '99'
+    ].join('|'),
   },
   WIDTH_HEIGHTS: {
     0: '0',