+++ packages/dialtone-vue2/components/tooltip/tooltip_default.story.vue
@@ -16,7 +16,6 @@
         :fallback-placements="$attrs.fallbackPlacements"
         :offset="$attrs.offset"
         :sticky="$attrs.sticky"
-        :append-to="$attrs.appendTo"
         :content-class="$attrs.contentClass"
         :transition="$attrs.transition"
         :show="$attrs.showTooltip"